Default scala version and java home are now being set.
If javaHome is not set in the properties, it uses the System.getProperty("java.home") version.

If scalaVersion is not set in the properties, it will look in the classpath for any org.scala-lang:scala-library:${version} value.  If none is found, it will default to 2.11.7.

There is still a check for scala or android plugins, but I hope to remove that shortly.