Updating for default versions for javaHome and scalaVersion.
All Java/Scala builds using Gradle are already running with a JDK.  If javaHome is not specified in the model, it should default to using the JDK that is running Gradle.

Scala version can be grabbed from the compile classpath.  For that matter, it should also be able to be retrieved from the testCompile or play configurations, as well (note that I still have not implemented Play suppor).  If scala-library is specified in any of those configurations and is *not* specified in the model, the version from the configuration will be used.

If no Scala version is specified (e.g., building a Java project rather than a Scala project), 2.11.7 is the default.